refactor: MealDetail 컴포넌트 분리
