+++ src/pages/Cafeteria/PCCafeteriaPage/components/PCMenuBlocks/PCMenuBlocks.module.scss
@@ -142,107 +142,3 @@
     line-height: 19.09px;
   }
 }
-
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0 0 0 / 40%);
-  z-index: 1;
-}
-
-.modal {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  z-index: 2;
-  width: 708px;
-  height: 459px;
-  padding: 18px;
-  gap: 12px;
-  border: none;
-  box-sizing: border-box;
-  border-radius: 12px;
-  background-color: white;
-  display: flex;
-  flex-direction: column;
-
-  &-header {
-    width: 100%;
-    height: 30px;
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-
-    &__place-chip {
-      display: flex;
-      flex-direction: row;
-      gap: 12px;
-    }
-
-    &__place {
-      font-family: Pretendard, sans-serif;
-      font-size: 20px;
-      font-weight: 500;
-      line-height: 26.25px;
-    }
-
-    &__chip {
-      width: auto;
-      height: 22px;
-      padding: 0 8px;
-      border-radius: 4px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      font-family: Pretendard, sans-serif;
-      font-size: 12px;
-      font-weight: 500;
-      line-height: 14.32px;
-
-      &--sold-out {
-        color: #ffad0d;
-        background-color: #fff7e1;
-      }
-
-      &--changed {
-        color: #4b4b4b;
-        background-color: #eee;
-      }
-    }
-
-    &__detail {
-      font-family: Pretendard, sans-serif;
-      font-size: 12px;
-      font-weight: 400;
-      line-height: 16.71px;
-      color: #8e8e8e;
-    }
-
-    &__close {
-      position: absolute;
-      top: 50%;
-      right: 0;
-      transform: translateY(-50%);
-      padding: 1px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      border: none;
-      background-color: transparent;
-      cursor: pointer;
-    }
-  }
-
-  &__image {
-    width: 672px;
-    height: calc(100% - 42.5px);
-    border-radius: 9px;
-    background-color: #f5f5f5;
-    object-fit: contain;
-  }
-}
